modify whole code for endpoints

+++ antarest/study/storage/variantstudy/model/command/remove_st_storage.py
@@ -55,12 +55,10 @@ class RemoveSTStorage(ICommand):
         storage = study_data.get_st_storage(self.area_id, self.storage_id)
 
         # Checks the storage is not referenced in any constraint
-        existing_constraints = study_data.get_st_storage_additional_constraints_for_area(self.area_id)
-        for constraint in existing_constraints:
-            if constraint.cluster == storage.id:
-                return command_failed(
-                    f"Short-term storage '{self.storage_id}' is referenced in the constraint '{constraint.id}'."
-                )
+        existing_constraints = study_data.get_st_storage_additional_constraints(self.area_id, self.storage_id)
+        if existing_constraints:
+            ids = {c.id for c in existing_constraints}
+            return command_failed(f"Short-term storage '{self.storage_id}' is referenced in the constraint(s) '{ids}'.")
 
         study_data.delete_st_storage(self.area_id, storage)
